testapp: add pidfile existence check before killing mupd8 and messageserver processes

+++ testapp.sh
@@ -24,12 +24,20 @@ start() {
 }
 
 stop() {
-    mupd8_pid=`perl -nw -e '/^(\d+)$/ and print $1;' $pidfile`
-    messageserver_pid=`perl -nw -e '/^(\d+)$/ and print $1;' $messageserver_pidfile`
-    echo "stopping the mupd8 app ($mupd8_pid)..."
-    kill $mupd8_pid
-    echo "stopping the messageserver ($messageserver_pid)..."
-    kill $messageserver_pid
+    if [ -f $pidfile ]; then
+      mupd8_pid=`perl -nw -e '/^(\d+)$/ and print $1;' $pidfile`
+      echo "stopping the mupd8 app ($mupd8_pid)..."
+      kill $mupd8_pid
+    else
+      echo "Can't find mupd8 process"
+    fi
+    if [ -f $messageserver_pidfile ]; then
+      messageserver_pid=`perl -nw -e '/^(\d+)$/ and print $1;' $messageserver_pidfile`
+      echo "stopping the messageserver ($messageserver_pid)..."
+      kill $messageserver_pid
+    else
+      echo "Can't find messageserver process"
+    fi
 }
 
 status() {
